refactor(core-weave): extract resource page model builders
